Move filesystem functions from thindev.rs to filesystem.rs
Removed thindev.rs - functionallity moved to devicemapper-rs.
